+++ nsdb-http/src/main/scala/io/radicalbit/nsdb/web/routes/QueryApi.scala
@@ -24,15 +24,15 @@ import akka.http.scaladsl.server.Directives._
 import akka.http.scaladsl.server.Route
 import akka.pattern.ask
 import akka.util.Timeout
-import com.fasterxml.jackson.annotation.{JsonSubTypes, JsonTypeInfo}
-import io.radicalbit.nsdb.common.{NSDbLongType, NSDbType}
-import io.radicalbit.nsdb.common.protocol.{Bit, NSDbSerializable}
+import io.radicalbit.nsdb.common.NSDbLongType
+import io.radicalbit.nsdb.common.protocol.Bit
 import io.radicalbit.nsdb.common.statement.{SQLStatement, SelectSQLStatement}
 import io.radicalbit.nsdb.protocol.MessageProtocol.Commands.ExecuteStatement
 import io.radicalbit.nsdb.protocol.MessageProtocol.Events._
 import io.radicalbit.nsdb.security.http.NSDBAuthProvider
 import io.radicalbit.nsdb.security.model.Metric
 import io.radicalbit.nsdb.sql.parser.StatementParserResult._
+import io.radicalbit.nsdb.web.Filters.Filter
 import io.radicalbit.nsdb.web.QueryEnriched
 import io.swagger.annotations._
 import javax.ws.rs.Path
@@ -41,56 +41,6 @@ import org.json4s.jackson.Serialization.write
 
 import scala.annotation.meta.field
 import scala.util.{Failure, Success}
-
-@ApiModel(description = "Filter Operators enumeration with [=, >, >=, <, <=, LIKE]")
-object FilterOperators extends Enumeration {
-  val Equality       = Value("=")
-  val GreaterThan    = Value(">")
-  val GreaterOrEqual = Value(">=")
-  val LessThan       = Value("<")
-  val LessOrEqual    = Value("<=")
-  val Like           = Value("LIKE")
-}
-
-@ApiModel(description = "Filter Nullability Operators enumeration with [ISNULL, ISNOTNULL]")
-object NullableOperators extends Enumeration {
-  val IsNull    = Value("ISNULL")
-  val IsNotNull = Value("ISNOTNULL")
-}
-
-@ApiModel(description = "Filter sealed trait", subTypes = Array(classOf[FilterNullableValue], classOf[FilterByValue]))
-@JsonTypeInfo(use = JsonTypeInfo.Id.NAME, property = "type")
-@JsonSubTypes(
-  Array(
-    new JsonSubTypes.Type(value = classOf[FilterByValue], name = "FilterByValue"),
-    new JsonSubTypes.Type(value = classOf[Generic], name = "FilterNullableValue")
-  ))
-sealed trait Filter extends NSDbSerializable
-
-case object Filter {
-  def unapply(arg: Filter): Option[(String, Option[NSDbType], String)] =
-    arg match {
-      case byValue: FilterByValue             => Some((byValue.dimension, Some(byValue.value), byValue.operator.toString))
-      case nullableValue: FilterNullableValue => Some((nullableValue.dimension, None, nullableValue.operator.toString))
-    }
-}
-
-@ApiModel(description = "Filter using operator", parent = classOf[Filter])
-case class FilterByValue(
-    @(ApiModelProperty @field)(value = "dimension on which apply condition") dimension: String,
-    @(ApiModelProperty @field)(value = "value of comparation") value: NSDbType,
-    @(ApiModelProperty @field)(
-      value = "filter comparison operator",
-      dataType = "io.radicalbit.nsdb.web.routes.FilterOperators") operator: FilterOperators.Value
-) extends Filter
-
-@ApiModel(description = "Filter for nullable", parent = classOf[Filter])
-case class FilterNullableValue(
-    @(ApiModelProperty @field)(value = "dimension on which apply condition") dimension: String,
-    @(ApiModelProperty @field)(
-      value = "filter nullability operator",
-      dataType = "io.radicalbit.nsdb.web.routes.NullableOperators") operator: NullableOperators.Value
-) extends Filter
 
 @ApiModel(description = "Query body")
 case class QueryBody(@(ApiModelProperty @field)(value = "database name") db: String,
@@ -147,7 +97,7 @@ trait QueryApi {
     ))
   def queryApi()(implicit logger: LoggingAdapter, format: Formats): Route = {
     path("query") {
-      post {
+      (post | get) {
         entity(as[QueryBody]) { qb =>
           optionalHeaderValueByName(authenticationProvider.headerName) { header =>
             authenticationProvider.authorizeMetric(ent = qb, header = header, writePermission = false) {
